combined protected data entry and audit trail
made a fish logging program, where the first instance of the program takes the username and password as the admin. can create a new account as a viewer. users can add fish caught with type of fish weight, length, where it was caught, and some notes. admin can delete entries and view database changes by timestamp and what was done by which user.